+++ NFlags/HelpPrinter.cs
@@ -1,0 +1,16 @@
+using NFlags.Commands;
+using NFlags.Utils;
+
+namespace NFlags
+{
+    /// <summary>
+    /// Prints help based on CommandConfig
+    /// </summary>
+    public static class HelpPrinter
+    {
+        /// <summary>
+        /// Default help printer implementation. 
+        /// </summary>
+        public static readonly IHelpPrinter Default = new DefaultHelpPrinter();
+    }
+}
