+++ NFlags/NFlagsConfig.cs
@@ -1,4 +1,5 @@
 using NFlags.TypeConverters;
+using NFlags.Utils;
 
 namespace NFlags
 {
@@ -16,9 +17,19 @@ namespace NFlags
         /// <param name="output">Output writing interface</param>
         /// <param name="environment">Environment variables provider</param>
         /// <param name="config">Configuration values provider</param>
+        /// <param name="helpPrinter">Help printer to generate help text</param>
         /// <param name="isExceptionHandlingEnabled">Is Exception handling enabled. Use exit code if enabled, otherwise throw exceptions. Default False</param>
         /// <param name="argumentConverters">List of param converters</param>
-        public NFlagsConfig(string name, string description, Dialect dialect, IOutput output, IEnvironment environment, IConfig config, bool isExceptionHandlingEnabled, IArgumentConverter[] argumentConverters)
+        public NFlagsConfig(
+            string name,
+            string description,
+            Dialect dialect,
+            IOutput output,
+            IEnvironment environment,
+            IConfig config,
+            IHelpPrinter helpPrinter,
+            bool isExceptionHandlingEnabled,
+            IArgumentConverter[] argumentConverters)
         {
             Name = name;
             Description = description;
@@ -27,6 +38,7 @@ namespace NFlags
             Output = output;
             Environment = environment;
             Config = config;
+            HelpPrinter = helpPrinter;
             ArgumentConverters = argumentConverters;
         }
 
@@ -69,5 +81,10 @@ namespace NFlags
         /// Configuration values provider
         /// </summary>
         public IConfig Config { get; }
+
+        /// <summary>
+        /// Help printer to generate help text
+        /// </summary>
+        public IHelpPrinter HelpPrinter { get; }
     }
 }
